Refactored auth buttons.
