+++ src/components/Mdx.js
@@ -4,7 +4,7 @@ import { MDXProvider } from "@mdx-js/react"
 import BrandCopy from "../components/BrandCopy"
 import { Link, ButtonLink } from "../components/Link"
 import Logo from "../components/Logo"
-import SignInButton from "../components/SignInButton"
+import { SignInButton } from "../components/AuthButtons"
 import StripeButton from "../components/StripeButton"
 import AngelsCheckout from "../components/AngelsCheckout"
 import AngelDonation from "../components/AngelDonation"
